+++ src/frontend/src/utils/styleUtils.ts
@@ -458,7 +458,6 @@ export const SIDEBAR_BUNDLES = [
   { display_name: "Perplexity", name: "perplexity", icon: "Perplexity" },
   { display_name: "pgvector", name: "pgvector", icon: "cpu" },
   { display_name: "Pinecone", name: "pinecone", icon: "Pinecone" },
-  { display_name: "Policies", name: "policies", icon: "shield-check" },
   { display_name: "Qdrant", name: "qdrant", icon: "Qdrant" },
   { display_name: "Redis", name: "redis", icon: "Redis" },
   { display_name: "SambaNova", name: "sambanova", icon: "SambaNova" },
