--- src/lfx/src/lfx/components/policies/__init__.py
@@ -1,34 +1,0 @@
-from __future__ import annotations
-
-from typing import TYPE_CHECKING, Any
-
-from lfx.components._importing import import_mod
-
-if TYPE_CHECKING:
-    from lfx.components.policies.policies_component import PoliciesComponent
-
-_dynamic_imports = {
-    "PoliciesComponent": "policies_component",
-}
-
-__all__ = [
-    "PoliciesComponent",
-]
-
-
-def __getattr__(attr_name: str) -> Any:
-    """Lazily import policies components on attribute access."""
-    if attr_name not in _dynamic_imports:
-        msg = f"module '{__name__}' has no attribute '{attr_name}'"
-        raise AttributeError(msg)
-    try:
-        result = import_mod(attr_name, _dynamic_imports[attr_name], __spec__.parent)
-    except (ModuleNotFoundError, ImportError, AttributeError) as e:
-        msg = f"Could not import '{attr_name}' from '{__name__}': {e}"
-        raise AttributeError(msg) from e
-    globals()[attr_name] = result
-    return result
-
-
-def __dir__() -> list[str]:
-    return list(__all__)
